requirements.txt
requests==2.31.0
google-generativeai==0.4.1
python-dotenv==1.0.1
colorama==0.4.6
